+++ test/utils/structs/EnumerableMap/UintToAddressMap.test.js
@@ -1,43 +1,19 @@
 const { BN, constants, expectEvent, expectRevert } = require('@openzeppelin/test-helpers');
 const { expect } = require('chai');
+const { expectMembersMatch } = require('./helpers');
 
-const zip = require('lodash.zip');
+const UintToAddressMapMock = artifacts.require('UintToAddressMapMock');
 
-const EnumerableMapMock = artifacts.require('EnumerableMapMock');
-
-contract('EnumerableMap', function (accounts) {
-  const [ accountA, accountB, accountC ] = accounts;
+contract('UintToAddressMap', function (accounts) {
+  const [accountA, accountB, accountC] = accounts;
 
   const keyA = new BN('7891');
   const keyB = new BN('451');
   const keyC = new BN('9592328');
 
   beforeEach(async function () {
-    this.map = await EnumerableMapMock.new();
+    this.map = await UintToAddressMapMock.new();
   });
-
-  async function expectMembersMatch (map, keys, values) {
-    expect(keys.length).to.equal(values.length);
-
-    await Promise.all(keys.map(async key =>
-      expect(await map.contains(key)).to.equal(true),
-    ));
-
-    expect(await map.length()).to.bignumber.equal(keys.length.toString());
-
-    expect(await Promise.all(keys.map(key =>
-      map.get(key),
-    ))).to.have.same.members(values);
-
-    // To compare key-value pairs, we zip keys and values, and convert BNs to
-    // strings to workaround Chai limitations when dealing with nested arrays
-    expect(await Promise.all([...Array(keys.length).keys()].map(async (index) => {
-      const entry = await map.at(index);
-      return [entry.key.toString(), entry.value];
-    }))).to.have.same.deep.members(
-      zip(keys.map(k => k.toString()), values),
-    );
-  }
 
   it('starts empty', async function () {
     expect(await this.map.contains(keyA)).to.equal(false);
@@ -64,7 +40,7 @@ contract('EnumerableMap', function (accounts) {
     it('returns false when adding keys already in the set', async function () {
       await this.map.set(keyA, accountA);
 
-      const receipt = (await this.map.set(keyA, accountA));
+      const receipt = await this.map.set(keyA, accountA);
       expectEvent(receipt, 'OperationResult', { result: false });
 
       await expectMembersMatch(this.map, [keyA], [accountA]);