+++ contracts/utils/Base64.sol
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: MIT
+// OpenZeppelin Contracts (last updated v4.5.0) (utils/Base64.sol)
 
 pragma solidity ^0.8.0;