+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+ * `ERC1155`: Add a `_afterTokenTransfer` hook for improved expendability. ([#3104](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/3104))
+
+## Unreleased
+
  * `ERC2891`: add implementation of the royalty standard, and the respective extensions for `ERC721` and `ERC1155`. ([#3012](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/3012))
  * `GovernorTimelockControl`: improve the `state()` function to have it reflect cases where a proposal has been canceled directly on the timelock. ([#2977](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/2977))
  * `Math`: add a `abs(int256)` method that returns the unsigned absolute value of a signed value. ([#2984](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/2984))