+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
- * `ERC1155`: Add a `_afterTokenTransfer` hook for improved expendability. ([#3104](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/3104))
+ * `ERC1155`: Add a `_afterTokenTransfer` hook for improved expendability. ([#3166](https://github.com/OpenZeppelin/openzeppelin-contracts/pull/3166))
 
 ## Unreleased